Added DB and Hibernate

+++ src/main/java/io/github/mat3e/HelloService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 class HelloService {
     static final String FALLBACK_NAME = "World";
-    static final Lang FALLBACK_LANG = new Lang(1L, "Hello", "en");
+    static final Lang FALLBACK_LANG = new Lang(1, "Hello", "en");
     private final Logger logger = LoggerFactory.getLogger(HelloService.class);
 
 
@@ -22,9 +22,9 @@ class HelloService {
     }
 
     String prepareGreeting(String name, String lang) {
-        Long langId;
+        Integer langId;
         try {
-            langId = Optional.ofNullable(lang).map(Long::valueOf).orElse(FALLBACK_LANG.getId());
+            langId = Optional.ofNullable(lang).map(Integer::valueOf).orElse(FALLBACK_LANG.getId());
         } catch (NumberFormatException c)
         {
             logger.warn("NON numeric language ID used " + lang);
